backend/requirements.txt
# Core dependencies for object detection
opencv-python>=4.8.0
ultralytics>=8.4.7
numpy>=1.21.0

# Anomaly detection dependencies (optional)
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
scikit-learn>=1.0.0

# Database and Storage dependencies
pymongo>=4.6.0
cloudinary>=1.36.0
python-dotenv>=1.0.0

# API Server dependencies
flask>=3.0.0
flask-cors>=4.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0

# AI/OpenRouter Integration
requests>=2.31.0

# Firebase
firebase-admin>=6.2.0

# Optional: For better performance
# torch-audio>=0.9.0  # If using audio features
# redis>=5.0.0  # For caching (optional)
# flask-jwt-extended>=4.5.0  # For authentication (optional)
# flask-limiter>=3.5.0  # For rate limiting (optional)
flask-socketio>=5.3.0  # For WebSocket support
python-socketio>=5.10.0
eventlet>=0.33.0
